# Vectorizing Gradient Descent — Multivariate Linear Regression and Python implementation
In this article, I shall go over the topic of arriving at the Vectorized Gradient-Descent formulae for the Cost function of the for Matrix…

![](https://cdn-images-1.medium.com/max/1200/1*lav3vWKSf34PvIFPiCaStg.jpeg)

In this article, I shall go over the topic of arriving at the **Vectorized Gradient-Descent formulae for the Cost function of the for Matrix form of training-data Equations.** And along with that the Fundamentals of Calculus (especially Partial Derivative) and Matrix Derivatives necessary to understand the process.

**So our target of this article is to understand the full Mathematics and the flow behind arriving at the below formulae**, which is the Vectorized Gradient of the training-data Matrix

![](https://cdn-images-1.medium.com/max/800/1*JPmxQrvKmjcAxiURYOCYUA.png)
undefined

### First a Refresher on basic Matrix Algebra

A matrix A over a field K or, simply, a matrix A (when K is implicit) is a rectangular array of scalars usually presented in the following form:

![](https://cdn-images-1.medium.com/max/800/1*Z7Y4gBTgQMym-MVR7DZ8pw.png)
undefined

The rows of such a matrix A are the m horizontal lists of scalars:

![](https://cdn-images-1.medium.com/max/800/1*Ev4S0SCYrgKd7SP7wXitzg.png)
undefined

and the columns of A are the n vertical lists of scalars:

![](https://cdn-images-1.medium.com/max/800/1*Hy4T3QGYGiT2YPwzoMdTSQ.png)
undefined

A matrix with m rows and n columns is called an m by n matrix, written m\*n. The pair of numbers m and n is called the size of the matrix. Two matrices A and B are equal, written A = B, if they have the same size and if corresponding elements are equal. Thus, the equality of two m \* n matrices is equivalent to a system of mn equalities, one for each corresponding pair of elements.
A matrix with only one row is called a row matrix or row vector, and a matrix with only one column is called a column matrix or column vector. A matrix whose entries are all zero is called a zero matrix and will usually be denoted by 0.

### Matrix Multiplication

The below image is taken from Khan Academy’s excellent linear algebra course.

![](https://cdn-images-1.medium.com/max/800/1*jYjtn1J2xM5iror3FXK4xQ.png)

**In above, each entry in the product matrix is the dot product of a row in the first matrix and a column in the second matrix**

**More explanation for higher dimension case** — If the product **AB = C** is defined, where C is denoted by \[cij\], then the
element cij is obtained by multiplying the elements in the ith row of **A** by the corresponding elements in the jth column of **B** and adding. Thus, if **A** has order k \* n, and **B** has order n \* p then

![](https://cdn-images-1.medium.com/max/800/1*fNTSrRCAcK_j4HvbLslqvA.png)
undefined

**then c11 is obtained by multiplying the elements in the first row of A by the corresponding elements in the first column of B and adding; hence,**

![](https://cdn-images-1.medium.com/max/800/1*7RSD_Z9KmiX5EKemHPL_ww.png)
undefined

**The element c12 is found by multiplying the elements in the first row of A by the corresponding elements in the second column of B and adding; hence,**

![](https://cdn-images-1.medium.com/max/800/1*UrH_UpFOZ6zZmv6BK8pJ4g.png)
undefined

**The element ckp below is obtained by multiplying the elements in the kth row of A by the corresponding elements in the pth column of B and adding; hence,**

![](https://cdn-images-1.medium.com/max/800/1*x779-OyJDcpwkLpunSLusw.png)
undefined

**There are four simple rules that will help us in multiplying matrices, listed here**

1\. Firstly, we can only multiply two matrices when the number of columns in
matrix A is equal to the number of rows in matrix B.

2\. Secondly, the first row of matrix A multiplied by the first column of matrix B
gives us the first element in the matrix AB, and so on.

3\. Thirdly, when multiplying, order matters — specifically, AB ≠ BA.

4\. Lastly, the element at row i, column j is the product of the ith row of matrix A and the jth column of matrix B.

The order in which we multiply matters. We must keep the matrices
in order, but we do have some flexibility. As we can see in the following equation, the parentheses can be moved:

![](https://cdn-images-1.medium.com/max/800/1*abGjdRm7NAazy9_-lmCqVg.png)
undefined

The following three rules also apply for Matrix Operation.

![](https://cdn-images-1.medium.com/max/800/1*ZIYKAV-1BlirjaAPxuKbSA.png)
undefined

Lets